Refactor component lib index
