+++ client/app/libs/components/index.jsx
@@ -1,3 +1,5 @@
+export { default as FeedContainer } from "./FeedContainer";
+export { default as Feed } from "./Feed";
 export { default as Flash } from "./Flash";
 export { default as Modal } from "./Modal";
 export { default as Progress } from "./Progress";
